+++ demo/core/domain/index.ts
@@ -2,6 +2,7 @@ export * from './result.js';
 export * from './errors.js';
 export * from './identity.js';
 export * from './authorization.js';
+export * from './password.js';
 export * from './slug.js';
 export * from './todo.js';
 export * from './card.js';